+++ src/quorum/command/network/delete.ts
@@ -1,18 +1,16 @@
-import { Arguments } from 'yargs'
 import config from '../../config'
 import Network from '../../service/network'
-import { logger, onCancel } from '../../../util'
+import { onCancel } from '../../../util'
 import prompts from 'prompts'
+import ora from 'ora'
 
 export const command = 'delete'
 
-export const desc = '刪除現有的 Quorum Network.'
+export const desc = '刪除現有的 Quorum Network'
 
 export const builder = {}
 
-export const handler = async (argv: Arguments) => {
-  logger.debug('exec network delete', argv.$0)
-
+export const handler = async () => {
   const network = new Network(config)
 
   let confirmDelete = true
@@ -20,14 +18,15 @@ export const handler = async (argv: Arguments) => {
   const response = await prompts({
     type: 'confirm',
     name: 'value',
-    message: '⚠️ The following processes will remove all existing files. Confirm to delete Quorum Network?',
+    message: '⚠️ The following processes will remove all network files. Confirm to delete Quorum Network?',
     initial: false,
   }, { onCancel })
 
   confirmDelete = response.value
 
   if (confirmDelete) {
+    const spinner = ora('Quorum Network Delete ...').start()
     await network.delete()
-    logger.info('Quorum Network delete Successfully!')
+    spinner.succeed('Quorum Network Delete Successfully!')
   }
 }