+++ src/quorum/command/network/down.ts
@@ -3,10 +3,11 @@ import config from '../../config'
 import Network from '../../service/network'
 import { logger, onCancel } from '../../../util'
 import prompts from 'prompts'
+import ora from 'ora'
 
 export const command = 'down'
 
-export const desc = '停止現有的 Quorum Network.'
+export const desc = '停止現有的 Quorum Network'
 
 export const builder = {}
 
@@ -27,7 +28,8 @@ export const handler = async (argv: Arguments) => {
   confirmDelete = response.value
 
   if (confirmDelete) {
+    const spinner = ora('Quorum Network Down ...').start()
     await network.down()
-    logger.info('Quorum Network down Successfully!')
+    spinner.succeed('Quorum Network Down Successfully!')
   }
 }