+++ src/quorum/instance/infra/docker/runner.ts
@@ -1,12 +1,12 @@
 import fs from 'fs'
 import stream from 'stream'
 import YAML from 'js-yaml'
-import { spawnSync } from 'child_process'
+import { spawn } from 'child_process'
 import Dockerode from 'dockerode'
 import { logger } from '../../../../util/logger'
 import { DockerCreateOptionsType, DockerStartOptionsType, DockerRunCommandType } from '../../../model/type/docker.type'
 import config from '../../../config'
-import { DockerError, FabricContainerError } from '../../../../util/error'
+import { DockerError, QuorumContainerError } from '../../../../util/error'
 import { DockerResultType, InfraRunner } from '../InfraRunner.interface'
 import { DockerComposeYamlInterface } from '../../../model/yaml/docker-compose/dockerComposeYaml'
 
@@ -27,15 +27,16 @@ export class Runner implements InfraRunner<DockerResultType> {
 
   public runCommand = async (payload: DockerRunCommandType): Promise<DockerResultType> => {
     await this.checkAndCreateNetwork(payload.network)
-    const { image, tag, commands } = payload
+    const { image, tag, commands, ignoreError } = payload
     const createOptions: DockerCreateOptionsType = payload.createOptions || {
+      AttachStdout: true,
       Env: (payload.envFile ? fs.readFileSync(payload.envFile, { encoding: 'utf8' }).toString().split(/\n|\r|\r\n/).filter((x) => /^\s*([\w.-]+)\s*=\s*(.*)?\s*$/.test(x)) : []).concat(payload.env || []),
       HostConfig: {
         AutoRemove: payload.autoRemove !== undefined ? payload.autoRemove : true,
         NetworkMode: payload.network,
         Binds: payload.volumes,
       },
-      User: `${config.UID}:${config.GID}`,
+      User: (payload.user !== false) ? `${config.UID}:${config.GID}` : '',
     }
     if (!this.existingImages[`${image}:${tag}`]) {
       const images = await this.dockerode.listImages({ filters: { reference: [`${image}:${tag}`] } })
@@ -67,12 +68,12 @@ export class Runner implements InfraRunner<DockerResultType> {
         startOptions)
       logger.silly(`run command output: \n${stdout}`)
       logger.debug(`docker run\n  image: ${image}\n  commands: ${commands.join(' ')}`)
-      if (dockerRunResult[0].StatusCode !== 0) {
-        throw new FabricContainerError(`[x] [in-docker-container error] ${stdout.split('\r\n').filter(x => x.match(/error/i) || stdout)}`, stdout)
+      if (dockerRunResult[0].StatusCode !== 0 && !ignoreError) {
+        throw new QuorumContainerError(`[x] [in-docker-container error] ${stdout.split('\r\n').filter(x => x.match(/error/i) || stdout)}`, stdout)
       }
-      return { statusCode: dockerRunResult[0].StatusCode, stdout }
+      return { statusCode: dockerRunResult[0].StatusCode, stdout: stdout.toString() }
     } catch (e: any) {
-      if (e instanceof FabricContainerError) { throw e }
+      if (e instanceof QuorumContainerError) { throw e }
       throw new DockerError(`[x] command [docker run]:${e.message}`)
     }
   }
@@ -87,7 +88,7 @@ export class Runner implements InfraRunner<DockerResultType> {
         NetworkMode: payload.network,
         Binds: payload.volumes,
       },
-      User: `${config.UID}:${config.GID}`,
+      User: (payload.user !== false) ? `${config.UID}:${config.GID}` : '',
     }
     const stdout = `docker command: \ndocker run -d -u ${config.UID}:${config.GID} ${createOptions.HostConfig?.NetworkMode ? `--network ${createOptions.HostConfig?.NetworkMode} ` : ''}${(createOptions.HostConfig?.Binds || []).map(x => `-v ${x} `).join('')}${(createOptions.Env || []).map(x => `--env ${x} `).join('')}${image}:${tag || 'latest'} ${commands.join(' ')}`
     logger.debug(stdout)
@@ -120,16 +121,19 @@ export class Runner implements InfraRunner<DockerResultType> {
   }
 
   // Docker Compose
-  private runSpawnSync (args: Array<string>): string {
-    logger.debug(`run spawnSync: docker-compose ${args.join(' ')}`)
-    const spawnReturn = spawnSync('docker-compose', [...args])
-    // TODO ! docker 裡面的 error 不能這樣抓
-    // TODO 如果 docker-compose 不存在不會報錯
-    if (spawnReturn.error) {
-      throw new DockerError(`[x] command [docker-compose]: ${spawnReturn.error.message}`)
-    }
-    logger.silly(spawnReturn.output.join('\n'))
-    return spawnReturn.output.join('\n')
+  private runSpawn (args: Array<string>):Promise<string> {
+    return new Promise((resolve) => {
+      logger.debug(`run spawnSync: docker-compose ${args.join(' ')}`)
+      const spawnReturn = spawn('docker-compose', [...args], { env: { ...process.env, UID: `${config.UID}`, GID: `${config.GID}` } })
+      // TODO ! docker 裡面的 error 不這樣抓
+      // TODO 如果 docker-compose 不存在不會報錯
+      spawnReturn.stdout.on('close', () => {
+        resolve(`docker-compose ${args.join(' ')} OK`)
+      })
+      spawnReturn.on('error', (error) => {
+        throw new DockerError(`[x] command [docker-compose]: ${error.message}`)
+      })
+    })
   }
 
   public upInBackground = async (dockerComposeFile: string) => {
@@ -139,7 +143,7 @@ export class Runner implements InfraRunner<DockerResultType> {
         await this.checkAndCreateNetwork(network)
       }
     }
-    return { stdout: this.runSpawnSync(['-f', dockerComposeFile, 'up', '-d']) }
+    return { stdout: await this.runSpawn(['-f', dockerComposeFile, 'up', '-d']) }
   }
 
   public upServiceInBackground = async (dockerComposeFile: string, service: string) => {
@@ -149,18 +153,18 @@ export class Runner implements InfraRunner<DockerResultType> {
         await this.checkAndCreateNetwork(network)
       }
     }
-    return { stdout: this.runSpawnSync(['-f', dockerComposeFile, 'up', '-d', '--', service]) }
+    return { stdout: await this.runSpawn(['-f', dockerComposeFile, 'up', '-d', '--', service]) }
   }
 
-  // eslint-disable-next-line require-await
   public downAndRemoveVolumes = async (dockerComposeFile: string) => {
-    // 為保留其他infra的操作空間，此method的type為(dockerComposeFile: string): Promise<InfraResultType>，雖然裡面沒有await，仍用async
-    return { stdout: this.runSpawnSync(['-f', dockerComposeFile, 'down', '--volumes']) }
+    return { stdout: await this.runSpawn(['-f', dockerComposeFile, 'down', '--volumes']) }
   }
 
-  // eslint-disable-next-line require-await
+  public downServiceAndRemoveVolumes = async (dockerComposeFile: string, service: string) => {
+    return { stdout: await this.runSpawn(['-f', dockerComposeFile, 'down', '--volumes', '--', service]) }
+  }
+
   public restart = async (dockerComposeFile: string, service: string[] = []) => {
-    // 為保留其他infra的操作空間，此method的type為(dockerComposeFile: string): Promise<InfraResultType>，雖然裡面沒有await，仍用async
-    return { stdout: this.runSpawnSync(['-f', dockerComposeFile, 'restart'].concat(service)) }
+    return { stdout: await this.runSpawn(['-f', dockerComposeFile, 'restart'].concat(service)) }
   }
 }