+++ src/quorum/instance/validator.ts
@@ -26,6 +26,11 @@ export default class Validator extends AbstractInstance {
     return await this.infra.downAndRemoveVolumes(this.dockerComposePath)
   }
 
+  public async downOneService (service: string): Promise<InfraRunnerResultType> {
+    logger.debug(`Validator instance ${service} down`)
+    return await this.infra.downServiceAndRemoveVolumes(this.dockerComposePath, service)
+  }
+
   public async restart (): Promise<InfraRunnerResultType> {
     logger.debug('Validator instance restart')
     return await this.infra.restart(this.dockerComposePath, [`explorer.${this.config.networkName}`])